app/requirements.txt
Django
gunicorn
psycopg2-binary
Pillow
djangorestframework
Flask==2.2.2
stripe==5.0.0
python-dotenv==0.21.0
pyjwt
Pillow
asgiref
attrs
autobahn
Automat
backports.zoneinfo
cffi
channels
constantly
cryptography
daphne
hyperlink
idna
incremental
pyasn1
pyasn1-modules
pycparser
pyOpenSSL
service-identity
six
sqlparse
Twisted
txaio
typing_extensions
tzdata
zope.interface
